requirements.txt
pycryptodome
requests
requests_toolbelt
aiohttp
revChatGPT
openai